refactor: registerIconSuggestion を移動
